+++ backend/apps/accounts/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth.decorators import user_passes_test
 
-from .modelss import User
+from .models import User
 
 admin.site.register(User)
