+++ backend/apps/accounts/forms/hospital_email.py
@@ -2,7 +2,7 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.db import transaction
 
-from apps.accounts.modelss import User
+from apps.accounts.models import User
 
 
 class HospitalEmailForm(forms.ModelForm):
