+++ backend/apps/accounts/src/statistics.py
@@ -1,7 +1,7 @@
 from django.db.models import Count
 from django.utils.translation import gettext_lazy as _
 
-from apps.accounts.modelss import Newsletter, User
+from apps.accounts.models import Newsletter, User
 from apps.iamstudent.forms import form_labels
 from apps.iamstudent.models import AUSBILDUNGS_TYPEN_COLUMNS, EmailToHospital, EmailToSend, Student
 from apps.ineedstudent.models import Hospital
