+++ backend/apps/iamstudent/models/student.py
@@ -5,7 +5,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from apps.accounts.modelss import User
+from apps.accounts.models import User
 from apps.mapview.files.map_data import plzs
