+++ backend/apps/ineedstudent/models/hospital.py
@@ -6,7 +6,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from apps.accounts.modelss import User
+from apps.accounts.models import User
 from apps.mapview.files.map_data import plzs
 
 # Create your models here.
